fix : 원본 url validation 로직 팩토리로 사용하기
